backend/requirements.txt
fastapi
uvicorn[standard]
langchain-openai
langchain-google-genai
langgraph
python-dotenv
requests
unidiff
aiohttp
tavily-python
google-generativeai